to_do_app/requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
Jinja2==3.1.6
kombu==5.5.0
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.2.0
prompt_toolkit==3.0.50
propcache==0.3.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
